backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator>=2.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic<2
python-socketio==5.15.0
python-engineio==4.11.0
Pillow==10.1.0
minio==7.2.5
